cloud_functions/fetch_historical/requirements.txt
functions-framework==3.*
google-cloud-storage==2.10.0
requests==2.31.0
pandas==2.1.0
datetime